stdout_mod can now return a blob
previous behavior: the result of stdout_mod was just used as is or
handed on to stdout_res

new behavior: if stdout_mod is given, then the result is deblobbed
before using it as output or handing it on to stdout_res

The new behavior was achievable before by using a dummy stdout_res
that returned args["stdout"] just to force deblobbing. This is a
somewhat strange and inconsistent behavior. The new behavior should be
intuitive and in the spirit of "everything is a blob".